refactor: move python package to subdir
