+++ gmail_chatbot/prompt_templates.py
