+++ tests/test_disk_store.py
@@ -16,7 +16,7 @@ from pathlib import Path
 import sys
 sys.path.insert(0, str(Path(__file__).parent.parent))
 
-from disk_store import DiskStore, DiskStoreError
+from gmail_chatbot.disk_store import DiskStore, DiskStoreError
 
 
 class TestDiskStore(unittest.TestCase):
