+++ tests/test_memory_models.py
@@ -13,7 +13,7 @@ from pathlib import Path
 import sys
 sys.path.insert(0, str(Path(__file__).parent.parent))
 
-from memory_models import MemoryEntry, MemoryKind, MemorySource
+from gmail_chatbot.memory_models import MemoryEntry, MemoryKind, MemorySource
 
 
 class TestMemoryEntry(unittest.TestCase):
